refactor: clean up test structure and fix syntax issues
- Remove unused WebDriver field from BaseTest and add proper cleanup
- Improve test method naming for better clarity
- Fix missing semicolon and typo in assertion messages
- Enhance overall code quality and maintainability
